requirements.txt
altair==4.2.0
anyascii==0.3.1
attrs==22.1.0
blinker==1.5
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
confection==0.0.3
contourpy==1.0.5
contractions==0.1.72
cycler==0.11.0
cymem==2.0.6
decorator==5.1.1
Distance==0.1.3
streamlit
sklearn
fuzzywuzzy
distance
bs4
entrypoints==0.4
fonttools==4.37.4
fuzzywuzzy==0.18.0
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
importlib-metadata==5.0.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.6.1
murmurhash==1.0.8
neattext==0.1.3
nltk==3.7
numpy==1.23.4
packaging==21.3
pandas==1.5.0
pathy==0.6.2
Pillow==9.2.0
plotly==5.10.0
preshed==3.0.7
protobuf==3.20.3
pyahocorasick==1.4.4
pyarrow==9.0.0
pydantic==1.9.2
pydeck==0.8.0b4
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
regex==2022.9.13
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
semver==2.13.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
streamlit==1.13.0
tenacity==8.1.0
textsearch==0.0.24
thinc==8.1.4
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wasabi==0.10.1
watchdog==2.1.9
wordcloud==1.8.2.2
zipp==3.9.0
sentence-transformers==2.2.2
sentencepiece==0.1.97



